# TicketReservation/src/main/resources/swagger.yaml
openapi: 3.0.0
info:
  title: Ticket Reservation API
  version: 1.0.0
servers:
  - url: http://localhost:8083/tickets  # Replace with your actual base URL
...
paths:
...
  /section/{section}:
    get:
      tags:
        - Ticket reservation    
      summary: Get Tickets by Section
      parameters:
        - in: path
          name: section
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of tickets for the section
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ticket'
        '404':
          description: No tickets found for the section
components:
  schemas:
    Ticket:
      type: object
      properties:
        ticketId:
          type: integer
        user:  # Assuming user is a separate object with its own properties
          type: object
        fromStation:
          type: string
        toStation:
          type: string
        price:
          type: number
        section:
          type: string
          enum:
            - Section_1  # Replace with actual enum values for section
            - Section_2
            # ...
        seat:
          type: integer
...